refactor modules and rename BytesReader into Reader
