+++ src/error.rs
@@ -27,7 +27,7 @@ pub type Result<T> = ::std::result::Result<T, Error>;
 ///
 /// - xml file when iterating xml `Event`s
 /// - element node when iterating on `Attribute`s. You can find xml
-/// relative position using `BytesReader::buffer_position()`
+/// relative position using `Reader::buffer_position()`
 pub type ResultPos<T> = ::std::result::Result<T, (Error, usize)>;
 
 impl fmt::Display for Error {
