+++ src/escape.rs
@@ -2,7 +2,7 @@
 
 use std::borrow::Cow;
 use error::{Error, ResultPos};
-use AsStr;
+use events::AsStr;
 
 // UTF-8 ranges and tags for encoding characters
 const TAG_CONT: u8 = 0b1000_0000;
